feat(suite-native): disabling of view only for device
