+++ suite-native/atoms/src/index.ts
@@ -1,5 +1,3 @@
-import exp from 'constants';
-
 export * from './AccordionItem';
 export * from './AlertBox';
 export * from './Text';
