+++ suite-native/intl/src/en.ts
@@ -449,7 +449,8 @@ export const en = {
         viewOnly: {
             title: 'View-only',
             emptyTitle: 'Connect your device to enable view-only',
-            subtitle: 'Check balances without connecting your Trezor. <about>See how it works</about>',
+            subtitle:
+                'Check balances without connecting your Trezor. <about>See how it works</about>',
             enableButton: 'Enable',
             about: {
                 title: 'View-only',
@@ -458,6 +459,19 @@ export const en = {
                 content:
                     '<li>Enable view-only to keep balances visible when your Trezor device is disconnected.</li><li>Your funds remain secure.</li><li>Your data remains private.</li><li>Stay updated on all transactions.</li><li>Create a new receive address.</li><li>To verify your receive address, simply reconnect your device.</li><li>Save time when assets are loading.</li>',
                 button: 'Got it',
+            },
+            toast: {
+                disabled: 'View-only disabled',
+                enabled: 'View-only enabled',
+            },
+            disableDialog: {
+                title: 'Disable view-only for {name}',
+                subtitle:
+                    'Your funds and transactions won’t be visible until you reconnect your device.',
+                buttons: {
+                    primary: 'Disable',
+                    secondary: 'Back',
+                },
             },
         },
     },
