+++ suite-native/module-settings/tsconfig.json
@@ -18,6 +18,7 @@
         {
             "path": "../../suite-common/wallet-core"
         },
+        { "path": "../alerts" },
         { "path": "../analytics" },
         { "path": "../atoms" },
         { "path": "../biometrics" },
